// emmaus_ui/src/app/services/models/impact-story-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ImpactStoryResponse {
  age?: number;
  beneficiaryName?: string;
  fullStory?: string;
  gender?: string;
  imageUrl?: string;
  location?: string;
  program?: string;
  programId?: string;
  publicId?: string;
  remark?: string;
  shortQuote?: string;
  status?: string;
  videoUrl?: string;
}